setup babel and package.module
this sets up babel to correctly compile for cross browser supprt and
allows compiling ES5 support for the same browsers.
using package.modules shoudl make it easie for npm to pick up the ES6
side of the package once this is supported correctly. users may have to
use the ES5 modules until all modules are converted to ES6 modules.