add builds to dist folder
Adds the midi and basic builds to the dist folder. This allows people to
import into node or the browser directly without being forced to use
ES6. We package the minified version incase someone is just wanting to
copy the file from the dist folder when using npm see #465.
We dont need to package these with version numbers as importing will be
more awkward when using npm.
We change package.main to point to the transpiled code to not force ES6
and we use package.module to point to the ES6 modules for bundlers to
find this or users can specify this location manually to use ES6.