always include page dimensions
when `switches.header_only` or `stop_on_warning` are passed to the parse
function page dimensions will not be included in the tune object. Even if
page directives are set in the header. This should not be the case as
if the directives are read the information should still be set in the
tune.