Implement a fix for race conditions during note loading
This fix addresses the issue with race conditions when multiple
instances attempt to load the same note. One instance will
overwrite the resolved AudioBuffer from another instance with the
'pending' string. This fix addresses this by setting an instance
of a Promise instead of strings that represent the status, and
modifies putNote() so that it makes no assumptions about the
status, and instead simply resolves the promise to obtain the
buffer.

See #721