ADD prettier
In the process of preparing the codemod that will operate the switch
to typescript, I'm having to add eslint and prettier: both of these
are going to be used by the codemod tool to format the code and detect
types and errors at compile time.
I do realize that Paul Rosen probably already has an eslintrc somewhere,
I'm not re-adding it here. Would it be useful to have it included in the
repo in the future?

add prettier ignore config