Fix metronome with changing meter
This is a workaround. Ideally for the following ABC snippet:
    M: 3/4
    %%MIDI drum ddd 76 77 77 90 90 90
first the meter change is handled, then the drum section.
Unfortunately as I see, the %%MIDI part is handled first,
so when the new drum definition has to be handled, the
meter is the previous one.

With this commit the drumDefinition is recalculated every
time when meter changes, so the updated drum pattern
will be used.

Also it will follow the standards when somebody just
forgot to update the drum pattern, and the meter changes.