Fix for crash on malformed @ positioned annotations
If a malformed "@"-style absolute positioned annotation was encountered, the renderer would throw an exception.

This change returns a reasonable chord value that that won't crash the rendering while a malformed absolute positioned annotation is being edited, which is a very common occurrence.

For example lets say user intends to add an annotation:

"@-30,20Hello"

While they are typing this, "@-30Hello" won't crash as it would before with this change in place.  

Once the user finally finished typing the full annotation: 

"@-30,20Hello"

then it will be rendered at the absolute position requested.